+++ tg_bot/modules/sql/warns_sql.py
@@ -1,5 +1,7 @@
 import threading
 
+from sqlalchemy.sql.sqltypes import BigInteger
+
 from tg_bot.modules.sql import BASE, SESSION
 from sqlalchemy import Boolean, Column, Integer, String, UnicodeText, distinct, func
 from sqlalchemy.dialects import postgresql
@@ -8,7 +10,7 @@ from sqlalchemy.dialects import postgresql
 class Warns(BASE):
     __tablename__ = "warns"
 
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
     chat_id = Column(String(14), primary_key=True)
     num_warns = Column(Integer, default=0)
     reasons = Column(postgresql.ARRAY(UnicodeText))
